STYDY_ANALYSER/studyanalyser/requirements.txt
Django==6.0.3
google-generativeai
python-dotenv
